Build Paella RPMs On Travis
This patch adjusts the Travis configuration to actually build Paella on
Travis to report possible problems. Additionally, RPMs are build for
each version and sent to pkg.opencast.org.

Note that for a successful upload both the variables PKGUPLOADUSER and
PKGUPLOADPASS need to be set in Travis. Although, tests will not fail if
they are not set.